smart-road-monitor/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
pymongo==4.5.0
dnspython==2.4.2
opencv-python==4.8.1.78
numpy==1.24.3
tensorflow==2.13.0
keras==2.13.1
Pillow==10.0.0
eventlet==0.33.3
geopy==2.3.0
python-socketio==5.9.0
requests==2.31.0
Werkzeug==2.3.7
APScheduler==3.10.4